+++ routes.php
@@ -2,20 +2,17 @@
 
 $router->get('/', 'HomeController@index');
 $router->get('/listings', 'ListingController@index');
-$router->get('/listings/create', 'ListingController@create');
+$router->get('/listings/create', 'ListingController@create', ['auth']);
+$router->get("/listings/edit/{id}", "ListingController@edit", ['auth']);
 $router->get("/listings/{id}", "ListingController@show");
-$router->get("/listings/edit/{id}", "ListingController@edit");
 
-$router->post('/listings', 'ListingController@store');
+$router->post('/listings', 'ListingController@store', ['auth']);
+$router->put("/listings/{id}", "ListingController@update", ['auth']);
+$router->delete('/listings/{id}', 'ListingController@destroy', ['auth']);
 
-$router->put("/listings/{id}", "ListingController@update");
+$router->get('/auth/register', 'UserController@create', ['guest']);
+$router->get('/auth/login', 'UserController@login', ['guest']);
 
-$router->delete('/listings/{id}', 'ListingController@destroy');
-
-$router->get('/auth/register', 'UserController@create');
-$router->get('/auth/login', 'UserController@login');
-
-$router->post('/auth/register', 'UserController@store');
-
-$router->post('/auth/logout', 'UserController@logout');
-$router->post('/auth/login', 'UserController@authenticate');
+$router->post('/auth/logout', 'UserController@logout', ['auth']);
+$router->post('/auth/register', 'UserController@store', ['guest']);
+$router->post('/auth/login', 'UserController@authenticate', ['guest']);
